+++ extensions/cornerstone/src/services/SegmentationService/SegmentationService.ts
@@ -55,6 +55,7 @@ export type SegmentationRepresentation = cstTypes.SegmentationRepresentation & {
   viewportId: string;
   id: string;
   label: string;
+  fallbackLabel?: string;
   styles: cstTypes.RepresentationStyle;
   segments: {
     [key: number]: SegmentRepresentation;
@@ -430,6 +431,7 @@ class SegmentationService extends PubSubService {
       },
       config: {
         label,
+        fallbackLabel: `S:${displaySet.SeriesNumber} ${displaySet.Modality}`,
         segments:
           options?.segments && Object.keys(options.segments).length > 0
             ? options.segments
@@ -583,6 +585,7 @@ class SegmentationService extends PubSubService {
       },
       config: {
         label: segDisplaySet.SeriesDescription,
+        fallbackLabel: `S:${segDisplaySet.SeriesNumber} ${segDisplaySet.Modality}`,
         segments,
       },
     };
@@ -664,6 +667,7 @@ class SegmentationService extends PubSubService {
       },
       config: {
         label: rtDisplaySet.SeriesDescription,
+        fallbackLabel: `S:${rtDisplaySet.SeriesNumber} ${rtDisplaySet.Modality}`,
       },
     };
 
@@ -1842,6 +1846,7 @@ class SegmentationService extends PubSubService {
       id: id,
       segmentationId,
       label: segmentation.label,
+      fallbackLabel: segmentation.fallbackLabel,
       active,
       type,
       visible,